obs-browser: merge upstream changes (#137)
* Remove unused code

* Disable system flash

This is the confirmed cause of some people crashing in recent CEF
versions, so just disable it all together for now.

* Rename getCurrentScene function for consistency

* Remove duplicate code for function execution

* Add replay buffer events to API

* Add function to save replay buffer

* Add JS events for source active/visible change

Partially fixes #72

* Add pause status & events to API

* Clean up & improve README

* Remove invalid cmake command

* Display errors in browser panels instead of an empty page

Co-authored-by: VodBox <dillon@vodbox.io>

* Inject CSS directly into a style element

Uses the JavaScript base64 decoder function atob().
This ensures proper newline and quote escaping.

* Use monospace font for Custom CSS property

* Set icon for browser source

* Size the browser window correctly on Mac HiDPI

* Unblacklist Intel igpu in single GPU configs

The blacklist of Intel igpu was motivated by issues with dual gpu setups
in laptops.  A few people with old Intel igpu (this was unfortunately
not documented) had also reported issues with hardware acceleration in
single gpu setups. Intel igpus were therefore blacklisted.  This however
prevents the hardware decoding of webm videos by Intel igpu in single
gpu setups.  With this commit hardware decoding of webm will be possible
for Intel igpu in single gpu setups while people who had issues with
hardware acceleration can still disable it.

* Fix compiler error in browser source when shared textures are disabled

* Enable linux support

* Use locale for Local file property

https://obsproject.com/mantis/view.php?id=1432

* Fix loading local files with a hash in the path

* Update browser version to 2.8.0

* Fix formatting

* Update version to 2.8.2

* Fix crash when replay buffer events are triggered

* Update browser version to 2.8.3

* Fix Chromium modifying process tokens/permissions

CEF allows Chromium to call HardenTokenIntegrityLevelPolicy, which
basically screws with our security descriptors, preventing our new
window capture method from working properly.  This is because even
though we already told CEF that we don't want sandboxing, it's basically
initializing it anyway for whatever reason.  So, instead of allowing it
to do that, just disable sandboxing directly.

Note that this is a massive hack.  But it's a really nice hack.

(I take full responsibility for this, this is just so much more clean
and simple than the alternatives.  -Jim)

Closes obsproject/obs-browser#204

* Don't use 512-sized array

It's pointless to use 512 here, it's a bit overkill seeing as the
original struct is only two pointers big.

Co-authored-by: Ryan Foster <RytoEX@gmail.com>
Co-authored-by: jp9000 <obs.jim@gmail.com>
Co-authored-by: Matt Gajownik <matt@wizardcm.com>
Co-authored-by: SCG82 <scg082+github@gmail.com>
Co-authored-by: VodBox <dillon@vodbox.io>
Co-authored-by: Clayton Groeneveld <claytong1214@gmail.com>
Co-authored-by: Theodore Dubois <tblodt@icloud.com>
Co-authored-by: pkv <pkv.stream@gmail.com>
Co-authored-by: Azat Khasanshin <azat.khasanshin@gmail.com>