Support fractional pixel ratios (#168)
* Don't inject CSS if the property is empty

This fixes an automatic breakpoint in Debug.

* Log CEF version (#250)

* Ignore LNK4099 warnings from libcef_dll_wrapper (#247)

* Update to new CEF audio API (#249)

This updates the audio functions due to API changes
in CEF & Chromium (AudioService API used for >= 4103).
Compatibility with CEF 3770 is kept for easier fallback.

The CEF browser subprocess is now loaded via absolute path,
as 4183 doesn't seem to parse relative path correctly.

Minor version of obs-browser updated to 2.9.0.

Co-authored-by: pkv <pkv@obsproject.com>
Co-authored-by: WizardCM <matt@obsproject.com>

* Disable Page Visibility API simulation on CEF 4183 and above

The issue seems to be fixed on CEF's side.
Feel free to revert if an edge case is found.

For reference, see e08909aaee094ac617a87c449118687891122162

* Fix Xcode linking issues introduced by CMake 3.19.0 (#248)

* Handle fractional scaling in browser panels

* Update CEF download URL in Readme (#253)

* Add new replay buffer save event to API

* Document obsExit event in the Readme

* Support fractional pixel ratios

* Use OBS_FRONTEND_EVENT_REPLAY_BUFFER_SAVED for OBS >= 26.1.x

Co-authored-by: Matt Gajownik <matt@obsproject.com>
Co-authored-by: pkv <pkv.stream@gmail.com>
Co-authored-by: RytoEX <RytoEX@gmail.com>
Co-authored-by: pkv <pkv@obsproject.com>
Co-authored-by: Patrick Heyer <PatTheMav@users.noreply.github.com>
Co-authored-by: Matt "Mookie" Thayer <m@mrmookie.com>