Handle cookie files from CEF 4638 on Windows [LT-524] (#196)
In case we failed to decrypt some cookie rows (there is an issue with
this in recent versions of CEF), delete the entire Cookies file before
it has a chance to be referenced by a `CefCookieManager`.

We must do this, since the version of CEF we bundle with the product at
this time differs from the version bundled with OBS Studio 27.2.x
(CEF 4638).

Any version of CEF, starting with 3770 and above, stores cookies in
an encrypted column.

CEF 3729, which we bundle with SE.Live does not support encrypted
cookies, so until this moment we used Win32 DPAPI to decrypt cookies
from CEF 3770, and store them in plain text for compatibility.

CEF 4638 bundled with OBS Studio 27.2.x no longer supports decryption
with DPAPI.
In addition, older versions of CEF treat Cookies files created
with OBS Studio 27.2.x as incompatible and refuse to load cookies from
those files.

This results in OBS Studio custom browser docks & native integrations
with streaming services such as Twitch to require the user of the
software to re-login each time OBS Studio is restarted.

Deleting the Cookies file will still require the end-user to re-login
once, yet, this is a lesser evil for the time being.